[#10] ImageType 위치 변경
