+++ core/common-ui/src/main/java/team/noweekend/core/common/ui/calendar/model/DateOfWeek.kt
@@ -2,7 +2,7 @@ package team.noweekend.core.common.ui.calendar.model
 
 import androidx.compose.runtime.Immutable
 import kotlinx.datetime.LocalDate
-import team.noweekend.core.common.ui.calendar.state.CalendarPagerState.ImageType
+import team.noweekend.core.common.ui.calendar.CalendarDataProvider.ImageType
 
 @Immutable
 data class DateOfWeek(
